Upgrade to Go 1.22, 1.23
Go 1.23 is out, so we can change supported versions to 1.22 and 1.23.
Switch CI to test against 1.22 and 1.23 only,
and merge test files that were previously split on Go versions
back together.
